rename: SvgIcon assets으로 폴더 구조 변경
